Individual contribution
In this project, I was responsible for the component of 3 model extensions’ report writing which included source extension, pumps and valves, as well as the preparing of presentation for this part. In the report writing, the formulations in mathematical representation based on SAS code were provided with detailed explanations. For each extension element' SAS result, which is before and after modification, were compared in order to detect the performance and effect of the module with extended constraints. I was also responsible for several references collection and wrote another version of the Introduction paragraph which associated with related literacy, but not be adopted at the final version of this report. Furthermore, the correlation between the extension result and the first case of cost-benefit analysis was also organized through exploring the pump efficiency in the presentation, as well as tanks volume change. In addition, the Github repository in terms of the whole page of this report is established to help each individual in this group to check the contribution of the other and make it accessible to edit the contents online directly.